backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI/ML
openai==1.10.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
tiktoken==0.5.2

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Git operations
gitpython==3.1.41

# Utilities
aiofiles==23.2.1
httpx==0.26.0
tenacity==8.2.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Linting
ruff==0.1.14
mypy==1.8.0
types-aiofiles==23.2.0.20240106
numpy<2
